return

+++ database/migrations/2020_11_16_063109_create_transaksi_sales_details_table.php
@@ -20,7 +20,7 @@ class CreateTransaksiSalesDetailsTable extends Migration
             $table->integer('stok_id');
             $table->integer('price');
             $table->integer('quantity')->nullable();
-            $table->double('diskon');
+            $table->integer('diskon');
             $table->integer('id_user')->nullable();
             $table->integer('harga_satuan')->nullable();
             $table->timestamps();
